+++ src/main/java/org/dinosaur/foodbowl/domain/auth/application/jwt/JwtConstant.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.auth.jwt;
+package org.dinosaur.foodbowl.domain.auth.application.jwt;
 
 public enum JwtConstant {
