+++ src/main/java/org/dinosaur/foodbowl/domain/auth/application/jwt/JwtTokenValid.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.auth.jwt;
+package org.dinosaur.foodbowl.domain.auth.application.jwt;
 
 public record JwtTokenValid(boolean isValid, String errorCode, String message) {
 }
